requirements.txt
torch==2.1.2
fair-esm==2.0.0
transformers==4.36.2
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
wandb==0.16.2
pytorch-lightning==2.1.3
biotite==0.38.0
biopython==1.81
# Development dependencies
black==23.12.1
isort==5.13.2
pytest==7.4.3
jupyter==1.0.0
